feat: add tool-call tracing and heartbeat events to agent bus
Emit AGENT_TOOL_CALL and AGENT_TOOL_RESULT events on the Redis bus for
every action executed inside agent_loop, surfacing tool activity that was
previously only visible in container stdout. Add a HEARTBEAT emit to
_think_loop so each agent pulses its queue depth and active task count on
the bus at every think_interval — confirming liveness during silent gaps.

Also fix Postgres parameter type inference error on NULL ttl_days by
casting to ::INTEGER, and copy core/ into the bridge Dockerfile.

+++ core/memory/long_term.py
@@ -449,7 +449,7 @@ class LongTermMemory:
                 """
                 INSERT INTO knowledge (agent, topic, title, content, tags, embedding, expires_at)
                 VALUES (%s, %s, %s, %s, %s, %s,
-                        CASE WHEN %s IS NOT NULL THEN NOW() + (%s || ' days')::INTERVAL ELSE NULL END)
+                        CASE WHEN %s::INTEGER IS NOT NULL THEN NOW() + (%s || ' days')::INTERVAL ELSE NULL END)
                 RETURNING id
             """,
                 (
@@ -489,7 +489,7 @@ class LongTermMemory:
                         """
                         INSERT INTO knowledge (agent, topic, title, content, tags, embedding, expires_at)
                         VALUES (%s, %s, %s, %s, %s, %s,
-                                CASE WHEN %s IS NOT NULL THEN NOW() + (%s || ' days')::INTERVAL ELSE NULL END)
+                                CASE WHEN %s::INTEGER IS NOT NULL THEN NOW() + (%s || ' days')::INTERVAL ELSE NULL END)
                         RETURNING id
                         """,
                         (
